Update ExtendedHydroGenerationCost.jl, ExtendedStorageCost.jl, & ExtendedRenewableCost.jl
Some minor but significant corrections. First of all, I put PSY wherever applicable and got rid of it, where not applicable (for instance, value in the getter function is a purely PowerLASCOPF object. It doesn't come from PSY). Secondly, I modified the type of the regularization_term as Union{T, Float64} instead of just T. This is because, whenever calling the constructor with nothing as argument, it initializes this field to 0.0 (which is of type Float64 and won't be possible to do so without it being of the Union{T, Float64} type.